Implemented branched pipeline based on doc retrieval
The are now two pipeline branches based on if the assistant was
able to retrieve documents similary to the query or not.
It will either base it's answer on documents and cite sources (if
docs were found) or give a general answer based on its training
data (in none were found).